requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-WTF==0.15.1
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-decouple==3.4
requests==2.24.0
urllib3==1.25.11
Werkzeug==1.0.1
WTForms==2.3.3
